+++ source/Common.h
@@ -300,6 +300,9 @@ typedef std::vector<BenchPathInfo> BenchPathInfoVec;
 #define XFER_STATS_ERRCOUNTLIST_ITEM_KIND		"k" // sparse encoding: error kind string
 #define XFER_STATS_ERRCOUNTLIST_ITEM_CNT		"c" // sparse encoding: number of errors of kind
 #define XFER_STATS_ERRCOUNT_STONEWALL			"ErrCountStoneWall"
+#define XFER_STATS_RETRYCOUNTLIST				"RetryCountList"
+#define XFER_STATS_RETRYCOUNTLIST_ITEM			"RetryCountList.item"
+#define XFER_STATS_RETRYWAITMILLIS				"RetryWaitMillis"
 #define XFER_STATS_CPUUTIL_STONEWALL			"CPUUtilStoneWall"
 #define XFER_STATS_CPUUTIL						"CPUUtil"